updated brads branch
changed int base to double base

+++ main.cpp
@@ -16,7 +16,7 @@ int main()
     return 0;
 }
 //--
-void power(int base, int exp)
+void power(double base, int exp)
 {
     int result = 1;
     for(int i = 0;i < exp;i++)
